Fundos/requirements.txt
pandas
supabase
python-dotenv
requests
